requirements.txt
aiohttp

beautifulsoup4

python-dotenv

fastapi
faiss-cpu

numpy

google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai

hf_xet
huggingface_hub[hf_xet]

Pillow

python-multipart
pydantic
docling
python-docx
PyYAML

sentence-transformers

torch --index-url https://download.pytorch.org/whl/cpu
transformers==4.50.3

uvicorn